chore: rename all apply related task files and introduce namespace
